- add guard for merchant and officer - layouting product detail - create error page (403, 404, 500) Next: - layout order page - layout review - layout analytics
Note/Suggestion:
- Need service/backend for handling merchant review average
