+++ backend/.env
@@ -4,8 +4,8 @@ TOKEN_SECRET = "hendryhendra"
 PORT=3000
 
 # Email settings
-MAIL_USERNAME= "af6f68be09de32"
-MAIL_PASSWORD= "ef1ee24164d602"
+MAIL_USERNAME= "156448c9e538b2"
+MAIL_PASSWORD= "4c639fe5e79c03"
 MAIL_HOST= "sandbox.smtp.mailtrap.io"
 MAIL_PORT=2525
 EMAIL_SENDER= "prstours@malaysia.gov.my"
