+++ package.json
@@ -25,6 +25,7 @@
     "@ngx-formly/material": "^6.2.2",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.7",
+    "@tinymce/tinymce-angular": "^7.0.0",
     "@types/swiper": "^6.0.0",
     "angular-datatables": "^16.0.1",
     "bcryptjs": "^2.4.3",
