+++ src/app/merchant/admin-merchant/layout/navbar-merchant/navbar-merchant.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { AuthService } from 'src/app/services/auth.service';
-import { TokenService } from 'src/app/token.service';
+import { TokenService } from 'src/app/services/token.service';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 import { LandingPageComponent } from 'src/app/landing-page/landing-page.component';
@@ -11,16 +11,17 @@ import { LandingPageComponent } from 'src/app/landing-page/landing-page.componen
   styleUrls: ['./navbar-merchant.component.css'],
 })
 export class NavbarMerchantComponent {
-
-  constructor(private router: Router, private authService: AuthService, private tokenService: TokenService) {
-  }
+  constructor(
+    private router: Router,
+    private authService: AuthService,
+    private tokenService: TokenService
+  ) {}
   dropdownOpen: boolean = false;
   toggleDropdown() {
     this.dropdownOpen = !this.dropdownOpen;
   }
 
-
-  onLogout(){
+  onLogout() {
     try {
       this.authService.logoutUser();
       this.router.navigateByUrl('/');
