+++ src/app/merchant/admin-merchant/product-add/product-add.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ProductService } from 'src/app/services/product.service';
 import { FormControl, FormGroup, Validators, FormArray } from '@angular/forms';
-import { TokenService } from 'src/app/token.service';
+import { TokenService } from 'src/app/services/token.service';
 import Swal from 'sweetalert2';
 
 @Component({
