+++ src/app/merchant/admin-merchant/product-update/product-update.component.html
@@ -116,17 +116,22 @@
                       >Product Description</label
                     >
                     <div class="mt-2">
-                      <textarea
-                        id="product-description"
-                        name="product-description"
-                        type="product-description"
-                        autocomplete="product-description"
-                        required
-                        formControlName="description"
-                        placeholder="Product Description"
+                      <editor
                         class="form-control bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5"
+                        apiKey="i16a5ggtgv4eg160h6zhazhvzq9qapp2o8zu96orza02lzdl"
+                        formControlName="description"
+                        [init]="{
+                          toolbar: 'undo redo | blocks fontfamily fontsize | bold italic underline strikethrough | link image media table mergetags | align lineheight | tinycomments | checklist numlist bullist indent outdent | emoticons charmap | removeformat',
+                          tinycomments_mode: 'embedded',
+                          tinycomments_author: 'Promo Tourism',
+                          browser_spellcheck: true,
+                          mergetags_list: [
+                            { value: 'First.Name', title: 'First Name' },
+                            { value: 'Email', title: 'Email' },
+                          ],
+                        }"
                       >
-                      </textarea>
+                      </editor>
                     </div>
                   </div>
                   <!-- Product Description: end -->
