--- src/app/order.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { OrderService } from './order.service';
-
-describe('OrderService', () => {
-  let service: OrderService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(OrderService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
