--- src/app/product/detail-product/detail-product.component.css
@@ -1,44 +1,0 @@
-.disabled-paypal {
-  pointer-events: none; /* Disable mouse events */
-  opacity: 0.5; /* Apply some visual indication of being disabled */
-}
-
-swiper-container {
-  width: 100%;
-  height: 300px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-swiper-slide {
-  background-size: cover;
-  background-position: center;
-}
-
-.mySwiper {
-  height: 90%;
-  width: 100%;
-}
-
-.mySwiper2 {
-  height: 20%;
-  box-sizing: border-box;
-  padding: 0px 0;
-}
-
-.mySwiper2 swiper-slide {
-  width: 100%;
-  height: 100%;
-  opacity: 1;
-}
-
-.mySwiper2 .swiper-slide-thumb-active {
-  opacity: 1;
-}
-
-swiper-slide img {
-  display: sw </swiper-slidek;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
