--- src/app/product/detail-product/detail-product.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DetailProductComponent } from './detail-product.component';
-
-describe('DetailProductComponent', () => {
-  let component: DetailProductComponent;
-  let fixture: ComponentFixture<DetailProductComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [DetailProductComponent]
-    });
-    fixture = TestBed.createComponent(DetailProductComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
