--- src/app/review.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ReviewService } from './review.service';
-
-describe('ReviewService', () => {
-  let service: ReviewService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ReviewService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
