+++ src/app/services/review.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ReviewService } from './review.service';
+
+describe('ReviewService', () => {
+  let service: ReviewService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ReviewService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
