--- src/app/token.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { TokenService } from './token.service';
-
-describe('TokenService', () => {
-  let service: TokenService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(TokenService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
