+++ src/app/widget/product-card/product-card.component.html
@@ -1,7 +1,7 @@
 <div
   class="bg-white border border-gray-200 rounded-2xl h-full min-w-48 max-w-full lg:hover:-translate-y-2 duration-300"
 >
-  <a [routerLink]="route" [skipLocationChange]="true" class="">
+  <a (click)="navigateToProductDetail(id)" class="">
     <div class="relative w-full h-44 rounded-2xl">
       <img
         class="w-full h-full object-cover object-center border-gray-200 rounded-2xl p-2"
